+++ transport/interface.go
@@ -23,15 +23,18 @@ package transport
 import (
 	"time"
 
+	"github.com/opentracing/opentracing-go"
 	"golang.org/x/net/context"
 )
 
 // Request is the fields used to make an RPC.
 type Request struct {
-	Method  string
-	Timeout time.Duration
-	Headers map[string]string
-	Body    []byte
+	Method   string
+	Timeout  time.Duration
+	Headers  map[string]string
+	Baggage  map[string]string
+	ShardKey string
+	Body     []byte
 }
 
 // Response represents the result of an RPC.
@@ -57,4 +60,5 @@ const (
 type Transport interface {
 	Call(ctx context.Context, request *Request) (*Response, error)
 	Protocol() Protocol
+	Tracer() opentracing.Tracer
 }