+++ transport/tchannel.go
@@ -28,6 +28,7 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/opentracing/opentracing-go"
 	"github.com/uber/tchannel-go"
 	"github.com/uber/tchannel-go/thrift"
 	"golang.org/x/net/context"
@@ -40,6 +41,7 @@ const rawHeadersKey = "_raw_"
 type tchan struct {
 	sc          *tchannel.SubChannel
 	callOptions *tchannel.CallOptions
+	tracer      opentracing.Tracer
 }
 
 // TChannelOptions are used to create a TChannel transport.
@@ -65,6 +67,10 @@ type TChannelOptions struct {
 	// TransportOpts are a list of options, mostly used to add or override
 	// TChannel's transport headers.
 	TransportOpts map[string]string
+
+	// Tracer is an instance of an opentracing tracer for baggage propagation
+	// and/or span submission.
+	Tracer opentracing.Tracer
 }
 
 // NewTChannel returns a Transport that calls a TChannel service.
@@ -88,6 +94,7 @@ func NewTChannel(opts TChannelOptions) (Transport, error) {
 	ch, err := tchannel.NewChannel(callerName, &tchannel.ChannelOptions{
 		Logger:      tchannel.NewLevelLogger(tchannel.SimpleLogger, level),
 		ProcessName: processName,
+		Tracer:      opts.Tracer,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to create TChannel: %v", err)
@@ -105,7 +112,12 @@ func NewTChannel(opts TChannelOptions) (Transport, error) {
 	return &tchan{
 		sc:          ch.GetSubChannel(opts.TargetService),
 		callOptions: callOpts,
+		tracer:      opts.Tracer,
 	}, nil
+}
+
+func (t *tchan) Tracer() opentracing.Tracer {
+	return t.tracer
 }
 
 func (t *tchan) Protocol() Protocol {
@@ -113,12 +125,23 @@ func (t *tchan) Protocol() Protocol {
 }
 
 func (t *tchan) Call(ctx context.Context, r *Request) (*Response, error) {
-	call, err := t.sc.BeginCall(ctx, r.Method, t.callOptions)
+	// We must create a shallow copy of the request headers because, at time of
+	// writing, we cannot prepare the trace headers before obtaining a TChannel
+	// call object. Consequently, we have to inject the headers and alter the
+	// request object for every call in a benchmark. Creating a shallow copy of
+	// the request object allows us to overwrite the headers reference without
+	// introducing a data race.
+	req := *r
+
+	call, err := t.sc.BeginCall(ctx, req.Method, t.callOptions)
+
 	if err != nil {
 		return nil, fmt.Errorf("begin call failed: %v", err)
 	}
 
-	if err := t.writeArgs(call, r); err != nil {
+	req.Headers = tchannel.InjectOutboundSpan(call.Response(), req.Headers)
+
+	if err := t.writeArgs(call, &req); err != nil {
 		return nil, err
 	}
 
@@ -127,8 +150,8 @@ func (t *tchan) Call(ctx context.Context, r *Request) (*Response, error) {
 		return nil, err
 	}
 
-	span := tchannel.CurrentSpan(ctx)
-	res.TransportFields["trace"] = fmt.Sprintf("%x", span.TraceID())
+	tchSpan := tchannel.CurrentSpan(ctx)
+	res.TransportFields["trace"] = fmt.Sprintf("%x", tchSpan.TraceID())
 	return res, nil
 }